pyproject.toml
[tool.poetry]
name = "teh-albina"
version = "0.1.0"
description = ""
authors = ["MaslovskiyNikita <nikitamaslovskiy999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
httpx = "^0.28.1"
python-dotenv = "^1.2.1"
django-environ = "^0.12.0"
gunicorn = "^25.0.2"
drf-spectacular = "^0.29.0"
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-mock = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
